Add verbiage about the soon-to-be deprecated components
The Creative SDK Image Editor UI, Color UI, Market Browser and Labs components are no longer supported so we're adding some verbiage to the sample projects and the user guides to inform everyone about the current state of these components and the corresponding user guides and sample projects. The guides and the projects will remain in this repository for now for reference purposes but existing code should be migrated away to stop supporting these components. No new code should be written that targets and uses these components. For more information please see https://creativesdk.zendesk.com/hc/en-us/articles/115004788463-End-of-Support-for-the-Creative-SDK-Image-Editor-UI-Color-UI-Market-Browser-and-Labs-Components